CL1299 - Bug: sometimes the Quill editor is outside the visible area
The quill editor introduces a toolbar (div) which is about 75px in height.
This throws off leafletjs which originally sized the popup to not include the toolbar.
When a popup is opened near the top of the viewscreen leafletjs tries to pan the map
to include the popup but is now unaware that the popup is larger and part of the popup
is opened out of the bounds of the viewscreen. To fix this problem a temporary spacer
is added when the popup is first created and in this step must now be removed.